day6/requirements.txt
python-dotenv>=1.0
openai>=1.40
numpy>=1.26
scikit-learn>=1.5
